refactor(FavoriteList): Extract the item in its own component
